started implementing concept in webcache
