+++ policies/cache_policies.cpp
@@ -1,214 +1,3 @@
-#include <iostream>
-#include <unordered_map>
-#include <map>
-#include <unordered_set>
-#include <vector>
-#include <queue>
-#include <list>
-#include <tuple>
-#include <assert.h>
-#include <random>
-
-using namespace std;
-
-typedef tuple<long, long> object_t; // objectid, size
-typedef list<object_t>::iterator list_iterator_t;
-typedef multimap<long double, object_t>::iterator map_iterator_t;
-
-// set to enable cache debugging
-//#define CDEBUG 1
-
-// util for debug
-#ifdef CDEBUG
-#define LOG(m,x,y,z) log_message(m,x,y,z)
-#else
-#define LOG(m,x,y,z)
-#endif
-
-void log_message(string m, double x, double y, double z) {
-  cerr << m << "," << x << "," << y  << "," << z << "\n";
-}
-
-/*
-  general cache class
-*/
-
-class Cache {
-protected:
-  // caching definitions
-  const long long cache_size;
-  long long current_size; // size of the cache
-  // hit statistics
-  long hits;
-  long long bytehits;
-
-  bool logStatistics;
-  virtual void hit(long size) {
-    if(logStatistics) {
-      hits++;
-      bytehits+=size;
-    }
-  }
-
-public:
-  Cache(long long cs) : cache_size(cs), current_size(0), hits(0), bytehits(0), logStatistics(false) {
-  }
-
-  virtual ~Cache(){};
-
-  virtual bool request (const long cur_req, const long long size) {return(false);}
-
-  // statistics
-  virtual void startStatistics() {
-    logStatistics = true;
-  }
-
-  virtual void stopStatistics() {
-    logStatistics = false;
-  }
-
-  virtual void resetStatistics() {
-    hits = 0;
-    bytehits = 0;
-  }
-  
-  virtual long getHits() const {
-    return(hits);
-  }
-
-  virtual bool getLogStatistics() const {
-    return(logStatistics);
-  }
-
-  virtual long long getBytehits() const {
-    return(bytehits);
-  }
-
-  virtual long long getCurrentSize() const {
-    return(current_size);
-  }
-
-  virtual long long getCacheSize() const {
-    return(cache_size);
-  }
-};
-
-/*
-  Least Recently Used implementation
-*/
-
-class LRUCache: public Cache {
-protected:
-  // ordered LRU list
-  list<object_t> cache_list;
-  // access to objects by storing respective iterators in unordered_map
-  unordered_map<long, list_iterator_t> cache_map;
-
-  virtual void hit(unordered_map<long, list_iterator_t>::const_iterator it, long size) {
-    cache_list.splice(cache_list.begin(), cache_list, it->second);
-    Cache::hit(size);
-  }
-
-  virtual void miss(const long cur_req, const long size) {
-    // object feasible to store?
-    if(size >= cache_size) {
-      LOG("error",0,size,cache_size);
-      return;
-    }
-    list_iterator_t lit;  
-    // check eviction needed
-    while (current_size + size > cache_size) {
-      // evict least popular (i.e. last element)
-      lit = cache_list.end();
-      lit--;
-      long esize = get<1>(*lit);
-      LOG("e",current_size,get<0>(*lit),esize);
-      current_size -= esize;
-      cache_map.erase(get<0>(*lit));
-      cache_list.erase(lit);
-    }
-    // admit new object
-    cache_list.push_front(object_t(cur_req, size)); 
-    cache_map[cur_req] = cache_list.begin();
-    current_size += size;
-    LOG("a",current_size,cur_req,size);
-  }
-
-
-public:
-  LRUCache(long long cs): Cache(cs) {}
-
-  LRUCache(const LRUCache& rhs): Cache(rhs.getCacheSize()) { /* copy constructor*/
-    this->cache_list = rhs.getCacheList();
-    for (list_iterator_t it = cache_list.begin(); it!=cache_list.end(); it++)
-      this->cache_map[get<0>(*it)] = it;
-    current_size = rhs.getCurrentSize();
-    hits = rhs.getHits();
-    bytehits = rhs.getBytehits();
-    logStatistics = rhs.getLogStatistics();
-    assert(this->cache_list.size() == this->cache_map.size());
-    //    cerr << "copied cache with " << cache_map.size() << " objects, " << double(current_size)/cache_size << " full" << endl;
-  }
-
-  ~LRUCache(){}
-
-  bool lookup (const long cur_req) const {
-
-    return(cache_map.count(cur_req)>0);
-  }
-
-  void evict (const long cur_req) {
-    if(lookup(cur_req)) {
-      list_iterator_t lit = cache_map[cur_req];
-      cache_map.erase(cur_req);
-      current_size -= get<1>(*lit);
-      cache_list.erase(lit);
-    }
-  }
-
-  bool request (const long cur_req, const long size) {
-    unordered_map<long, list_iterator_t>::const_iterator it;
-    it = cache_map.find(cur_req);
-    if(it != cache_map.end())
-      {
-	if (size == get<1>(*(it->second))) {
-	  // hit and consistent size
-	  LOG("h",0,cur_req,size);
-	  hit(it, size);
-	  return(true);
-	}
-	else { // inconsistent size -> treat as miss and delete inconsistent entry
-	  evict(cur_req);
-	  //	  cerr << "deleted outdated object" << endl;
-	}
-      }
-    miss(cur_req, size);
-    return(false);
-  }
-
-  const list<object_t> getCacheList() const {
-    return (cache_list);
-  }
-  const unordered_map<long, list_iterator_t> getCacheMap() const {
-    return (cache_map);
-  }
-};
-
-
-
-/*
-  FIFO
-*/
-
-class FIFOCache: public LRUCache {
-protected:
-  virtual void hit(unordered_map<long, list_iterator_t>::const_iterator it, long size) {
-    Cache::hit(size);
-  }
-public:
-  FIFOCache(long long cs): LRUCache(cs) {}
-  ~FIFOCache(){}
-};
